+++ db/migrate/20150214222732_create_statistics.rb
@@ -3,8 +3,8 @@ class CreateStatistics < ActiveRecord::Migration
     create_table :statistics do |t|
       t.integer :type_id
       t.datetime :timestamp
-      t.integer :system_id
-      t.integer :service_id
+      t.bigint :system_id
+      t.bigint :service_id
       t.float :stat
 
       t.timestamps
